# Predictive Analytics Pipeline on Kubernetes/api/src/main/java/com/portfolio/pipeline/api/model/PredictionResponse.java
package com.portfolio.pipeline.api.model;

import java.time.Instant;

public record PredictionResponse(
    String eventType,
    Instant windowStart,
    Instant windowEnd,
    long eventCount,
    double avgValue,
    double predictionScore
) {}